Reusing Blas Level 1 and Lapack(Givens Rotations) kernels in GMRES Solver
1. Blas and Lapack libraries are linked as part of sparse's cmake build
   system to reuse kernels wherever applicable
2. norm2, scale, dot and axpby level-1 kernels from Blas are reused in
   aocl-sparse's GMRES
3. Lapack's plane rotation routines dlartg and slartg are used to
   perform Givens rotations in GMRES Solver

AMD-Internal: CPUPL-2546

+++ tests/examples/sample_itsol_d_gmres.cpp
@@ -189,7 +189,7 @@ int main()
         printf("Warning gmres iteration limit option could not be set, exit status = %d\n",
                exit_status);
 
-    sprintf(rs_iters_string, "%d", rs_iters);
+    sprintf(rs_iters_string, "%d", (int)rs_iters);
     exit_status = aoclsparse_itsol_option_set(handle, "gmres restart iterations", rs_iters_string);
     if(exit_status != aoclsparse_status_success)
         printf("Warning gmres restart iterations option could not be set, exit status = %d\n",
@@ -207,11 +207,11 @@ int main()
         norm = calculate_l2Norm_solvers(expected_sol, x, n);
         printf("input = %s\n", filename.c_str());
         printf("solver = %s\n", gmres_sol);
-        printf("no of rows = %d\n", m);
-        printf("no of cols = %d\n", n);
-        printf("no of nnz = %d\n", nnz);
+        printf("no of rows = %d\n", (int)m);
+        printf("no of cols = %d\n", (int)n);
+        printf("no of nnz = %d\n", (int)nnz);
         printf("monitoring tolerance = %e\n", tol);
-        printf("restart iterations = %d\n", rs_iters);
+        printf("restart iterations = %d\n", (int)rs_iters);
         printf("residual achieved = %e\n", rinfo[0]);
         printf("total iterations = %d\n", (int)rinfo[30]);
         printf("l2 Norm = %e\n", norm);
